pyproject.toml
[project]
name = "stock-trading-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "yfinance>=0.2.50",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "numpy>=2.0.0",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "torch>=2.5.0",
    "transformers>=4.47.0",
    "sentence-transformers>=3.3.0",
    "prophet>=1.1.6",
    "praw>=7.8.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "finnhub-python>=2.4.20",
    "pgvector>=0.3.6",
    "feedparser>=6.0.11",
    "newsapi-python>=0.2.7",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
